Decorator started, BuildingFacotry updated, Building updated
I have started the implement the Decorator class, I have added a getCost method to the Building base class and started to include these functions in the derived Building classes. I have also added the amount of resources needed per resource in the BuildingFactory. So then before the Building is constructed the BuildingFactory requests these resources from the ResourceManager and if there is enough the Building is constructed otherwise null will be returned, as well as an error message.